[Messaging] Fix Functions test host credential use
The focus of these changes is to fix the approach used for token credential
management in the Function extensions end-to-end tests for Event Hubs and Service
Bus.  The previous approach used simple logic to set in-memory configuration with
assumptions about the test environment.  These assumptions are no longer accurate
with the security changes in the testing pipelines.

The new approach delegates to the test environment for credential management, as
it has the conditional logic for environmental factors needed for credential
selection and creation.  The configuration-to-credential mapping and creation
logic is part of the `Microsoft.Extensions.Azure` package and has full test
coverage there; there is no benefit to duplicating test environment logic just to
force Function configuration to control the credential locally for the test host.